MATH: created new clustering module and moved several classes from legacy
