+++ commons-math-clustering/src/main/java/org/apache/commons/math4/clustering/Cluster.java
@@ -15,10 +15,11 @@
  * limitations under the License.
  */
 
-package org.apache.commons.math4.legacy.ml.clustering;
+package org.apache.commons.math4.clustering;
 
 import java.util.ArrayList;
 import java.util.List;
+import org.apache.commons.math4.clustering.Clusterable;
 
 /**
  * Cluster holding a set of {@link Clusterable} points.
