+++ commons-math-legacy/pom.xml
@@ -132,6 +132,12 @@
       <artifactId>commons-geometry-euclidean</artifactId>
       <scope>test</scope>
     </dependency>
+    <dependency>
+      <groupId>org.apache.commons</groupId>
+      <artifactId>commons-math4-clustering</artifactId>
+      <version>4.0-SNAPSHOT</version>
+      <scope>compile</scope>
+    </dependency>
 
   </dependencies>
