+++ commons-math-legacy/src/test/java/org/apache/commons/math4/legacy/ml/clustering/MultiKMeansPlusPlusClustererTest.java
@@ -20,7 +20,8 @@ package org.apache.commons.math4.legacy.ml.clustering;
 
 import java.util.Arrays;
 import java.util.List;
-
+import org.apache.commons.math4.clustering.Clusterable;
+import org.apache.commons.math4.clustering.DoublePoint;
 import org.junit.Assert;
 import org.junit.Test;
